fastapi/pyproject.toml
[tool.poetry]
name = "rinha-de-backend-2023-q3-vncsna-fastapi"
version = "1.0.0"
description = "Proposta (atrasada) pra Rinha de Backend 2023Q3 em FastAPI"
authors = ["Vinicius <vncsna@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["pydantic-settings", "uvicorn"], version = "^0.109.2"}
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
